basic equities and bond logic works for all cycles

+++ calculator/forms.py
@@ -10,7 +10,7 @@ class CalculatorForm(forms.Form):
         ('single_simulation', 'Single Simulation Cycle'),
     )
     data_method = forms.ChoiceField(choices=DATA_METHOD_CHOICES)
-    portfolio_value = forms.IntegerField(initial="1000000")
+    portfolio_value = forms.DecimalField(initial="1000000")
     equities = forms.IntegerField(initial="75")
     bonds = forms.IntegerField(initial="25")
     gold = forms.IntegerField(initial="0")
@@ -21,10 +21,10 @@ class CalculatorForm(forms.Form):
         ('inflation_adjusted', 'Inflation Adjusted'),
     )
     spending_plan = forms.ChoiceField(choices=SPENDING_PLAN_CHOICES)
-    initial_yearly_spending = forms.IntegerField(initial="40000")
-    ss_annual_value = forms.IntegerField(initial="0")
+    initial_yearly_spending = forms.DecimalField(initial="40000")
+    ss_annual_value = forms.DecimalField(initial="0")
     ss_start_year = forms.IntegerField(initial="2032")
     ss_end_year = forms.IntegerField(initial="2100")
-    ss_spouse_annual_value = forms.IntegerField(initial="0")
+    ss_spouse_annual_value = forms.DecimalField(initial="0")
     ss_spouse_start_year = forms.IntegerField(initial="2032")
     ss_spouse_end_year = forms.IntegerField(initial="2100")
